requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
cmake==3.27.5
contourpy==1.1.1
cycler==0.11.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
lit==17.0.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.1
protobuf==4.24.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.6.1.post1
tensorrt-bindings==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
timm==0.9.7
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==2.3.7
wrapt==1.14.1
